ai-vectorsearch-chatbot/src/a-virtual-machine/www/autonomous.database.23ai.generative.ai/requirements.txt
oracle-ads
Faker==25.2.0
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.33
oracledb==2.3.0
pandas==2.2.2
python-dotenv==1.0.1
streamlit==1.37.1